[SubtasksJS-6] Убрала из функции лишнее
